Refactor fractal calculations into modular structure
